Major Fixes check description
- Toast UI has been revamped
- Added the clear function to the clear button that would clear the references field in testrail as well once clear has been clicked
- Broke down the Page to its components, styles, actions, and views
