+++ src/components/corsHeaders.js
@@ -1,0 +1,11 @@
+export const corsHeaders = {
+  "Access-Control-Allow-Origin": "https://trajector1.aha.io",
+  "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, OPTIONS",
+  "Access-Control-Allow-Headers": "*",
+};
+
+export const corsHeadersWithAuth = {
+  "Access-Control-Allow-Origin": "https://trajector1.aha.io",
+  "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, OPTIONS",
+  "Access-Control-Allow-Headers": "Content-Type, Authorization",
+};
