+++ src/styles/TestAhaStyles.js
@@ -1,0 +1,115 @@
+export const TestAhaStyles = {
+  select_style: {
+    width: "90%",
+    marginRight: "2.5%",
+  },
+  select_style_with_center: {
+    display: "flex",
+    flexDirection: "row",
+    alignItems: "center",
+  },
+
+  input_style: {
+    width: "25%",
+    marginRight: "2.5%",
+    paddingLeft: "10px",
+    display: "none",
+  },
+
+  loading_style: {
+    width: "12px",
+    height: "12px",
+    borderRadius: "50%",
+    border: "4px solid rgba(0, 0, 0, 0.1)",
+    borderLeftColor: "#7986CB",
+    animation: "spin 1s linear infinite",
+  },
+
+  submit_button: {
+    width: "18.5%",
+    backgroundColor: "rgb(0, 115, 207)",
+    border: "none",
+    borderWidth: "1px",
+    color: "white",
+    borderRadius: ".3rem",
+    boxSizing: "border-box",
+    cursor: "pointer",
+    display: "inline-block",
+    fontSize: ".8125rem",
+    lineHeight: "1.25rem",
+    minWidth: "2rem",
+    outline: "none!important",
+    padding: "4px 12px",
+    textAlign: "center",
+    textDecoration: "none",
+    textShadow: "none",
+    WebkitUserSelect: "none",
+    userSelect: "none",
+    verticalAlign: "middle",
+    marginTop: "10px",
+    marginBottom: "10px",
+    marginRight: "10px",
+  },
+
+  load_submit_spinner: {
+    width: "10px",
+    height: "10px",
+    borderRadius: "50%",
+    border: "4px solid rgba(0, 0, 0, 0.1)",
+    borderLeftColor: "#7986CB",
+    animation: "spin 1s linear infinite",
+  },
+
+  tcl_div: {
+    display: "flex",
+    flexDirection: "row",
+    gap: 3,
+    marginTop: "10px",
+    justifyContent: "flex-end",
+  },
+  Add_TestCase_Button: {
+    backgroundColor: "rgb(0,115,207)",
+    borderWidth: "1px",
+    color: "black",
+    border: "1px solid black",
+    borderRadius: ".3rem",
+    boxSizing: "border-box",
+    cursor: "pointer",
+    display: "inline-block",
+    fontSize: ".8125rem",
+    lineHeight: "1.25rem",
+    minWidth: "2rem",
+    outline: "none!important",
+    padding: "2px 12px",
+    textAlign: "center",
+    textDecoration: "none",
+    textShadow: "none",
+    WebkitUserSelect: "none",
+    userSelect: "none",
+    position: "relative",
+    top: "-1.px",
+  },
+  Clear_Button: {
+    width: "10%",
+    backgroundColor: "white",
+    border: "1px solid black ",
+    borderWidth: "1px",
+    color: "black",
+    borderRadius: ".3rem",
+    boxSizing: "border-box",
+    cursor: "pointer",
+    display: "inline-block",
+    fontSize: ".8125rem",
+    lineHeight: "1.25rem",
+    outline: "none!important",
+    padding: "2px 12px",
+    textAlign: "center",
+    textDecoration: "none",
+    textShadow: "none",
+    WebkitUserSelect: "none",
+    userSelect: "none",
+    position: "relative",
+    minWidth: "58px",
+    maxWidth: "58px",
+  },
+};
