+++ src/styles/TestDashStyles.js
@@ -1,0 +1,99 @@
+export const TestDashStyles = {
+  parent_div: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  loading_div: {
+    width: "12px",
+    height: "12px",
+    borderRadius: "50%",
+    border: "4px solid rgba(0, 0, 0, 0.1)",
+    borderLeftColor: "#7986CB",
+    animation: "spin 1s linear infinite",
+    zIndex: "10",
+    marginBottom: "20px",
+  },
+  divider: { display: "flex", width: "100%", borderTop: "1px solid black" },
+  latest_test_runs: {
+    flex: "flex-row",
+    marginTop: "10px",
+    border: "1px solid #ECECEC ",
+    padding: "10px 10px",
+    borderRadius: "10px",
+    backgroundColor: "#ECECEC",
+  },
+  h2_div_1: {
+    display: "flex",
+    flexDirection: "row",
+    alignItems: "center",
+  },
+  header_div_2: {
+    width: "12px",
+    height: "12px",
+    marginRight: "10px",
+  },
+
+  remLoading_style: {
+    width: "12px",
+    height: "12px",
+    borderRadius: "50%",
+    border: "4px solid rgba(0, 0, 0, 0.1)",
+    borderLeftColor: "#7986CB",
+    animation: "spin 1s linear infinite",
+    zIndex: "10",
+  },
+
+  strong_text_header_with_margin: {
+    fontSize: "1.3rem",
+    marginLeft: "40px",
+  },
+  tempFTC_Map_li: { listStyle: "none", marginBottom: ".5rem" },
+
+  tempFTC_Map_div: {
+    display: "flex",
+    flexDirection: "row",
+    gap: "3rem",
+    marginLeft: "-1rem",
+  },
+  tempFTC_Map_button: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    border: "1px solid black",
+    borderRadius: "30%",
+    maxHeight: "30px",
+    maxWidth: "30px",
+    backgroundColor: "transparent",
+    cursor: "pointer",
+  },
+
+  tempFTC_div_2: {
+    marginTop: "10px",
+    border: "1px solid #ECECEC",
+    padding: "10px",
+    borderRadius: "10px",
+    backgroundColor: "#ECECEC",
+  },
+
+  tempFTC_h2_2: {
+    display: "flex",
+    flexDirection: "row",
+    alignItems: "center",
+  },
+
+  tempFTC_div_3: {
+    width: "12px",
+    height: "12px",
+    marginRight: "10px",
+  },
+  addLoading: {
+    width: "12px",
+    height: "12px",
+    borderRadius: "50%",
+    border: "4px solid rgba(0, 0, 0, 0.1)",
+    borderLeftColor: "#7986CB",
+    animation: "spin 1s linear infinite",
+    zIndex: "10",
+  },
+};
